Clarify loading from own device
If player has copied the game files to a device (e.g. tablet computer)
and loads them with a web browser from the device (not from a web
server), then it should not be possible to play over the Internet.

Detect the application protocol used to request the game files. If it
does not start with `http', disable and `gray out' the play through
Internet button.

Update the game description accordingly.