Detect set up web server but not game server
If player loads the game files from a web server but the game server
for two players over the Internet is not set up and player pushes the
Play button, the game mode hangs on "Attempting to start..." message.

Detect this case and inform about it the player to avoid causing the
confusion and disable the button from being pushed again without
reloading the game web page.